SHT15libmod2
This is an Arduino library for Sensirion's SHT15 Humidity sensor. This
library performs bit-banging for data retrieval, and exploits all
features of the SHT15 mentioned in V5 of the datasheet, including
temperature compensation of relative humidity. This library utilizes an
external temperature sensor's value to compensate for relative humidity
for best performance.